重构结算机制

+++ Assets/Scripts/Logic/EventSystem/PTrigger.cs
@@ -20,7 +20,7 @@ public class PTrigger : PObject{
     public Action<PGame> Effect = DefaultAction;
     public Predicate<PGame> Condition = DefaultCondition;
     public Predicate<PGame> AICondition = DefaultCondition;
-    public float AIPriority = 1; // 权重大的效果AI优先发动
+    public int AIPriority = 1; // 权重大的效果AI优先发动
     public bool CanRepeat = false; // 是否可以被重复发动
 
     public delegate bool PlayerCondition(PGame Game, PPlayer Player);
